move providers into constants
