--- app/components/Web3Auth/providers.js
@@ -1,93 +1,0 @@
-const google = require('assets/providers/google.png');
-const facebook = require('assets/providers/facebook.png');
-const apple = require('assets/providers/apple.png');
-const discord = require('assets/providers/discord.png');
-const email = require('assets/providers/email.png');
-const github = require('assets/providers/github.png');
-const jwt = require('assets/providers/jwt.png');
-const kakao = require('assets/providers/kakao.png');
-const line = require('assets/providers/line.png');
-const linkedin = require('assets/providers/linkedin.png');
-const reddit = require('assets/providers/reddit.png');
-const twitch = require('assets/providers/twitch.png');
-const twitter = require('assets/providers/twitter.png');
-const weibo = require('assets/providers/weibo.png');
-const weixin = require('assets/providers/weixin.png');
-
-export const providers = [
-	{
-		id: 1,
-		name: 'google',
-		image: google,
-	},
-	{
-		id: 2,
-		name: 'facebook',
-		image: facebook,
-	},
-	// {
-	// 	id: 3,
-	// 	name: 'reddit',
-	// 	image: reddit,
-	// },
-	{
-		id: 4,
-		name: 'discord',
-		image: discord,
-	},
-	// {
-	// 	id: 5,
-	// 	name: 'twitch',
-	// 	image: twitch,
-	// },
-	{
-		id: 6,
-		name: 'apple',
-		image: apple,
-	},
-	// {
-	// 	id: 7,
-	// 	name: 'line',
-	// 	image: line,
-	// },
-	{
-		id: 8,
-		name: 'github',
-		image: github,
-	},
-	{
-		id: 9,
-		name: 'kakao',
-		image: kakao,
-	},
-	{
-		id: 10,
-		name: 'linkedin',
-		image: linkedin,
-	},
-	{
-		id: 11,
-		name: 'twitter',
-		image: twitter,
-	},
-	// {
-	// 	id: 12,
-	// 	name: 'weibo',
-	// 	image: weibo,
-	// },
-	// {
-	// 	id: 13,
-	// 	name: 'wechat',
-	// 	image: weixin,
-	// },
-	// {
-	// 	id: 14,
-	// 	name: 'email_passwordless',
-	// 	image: email,
-	// },
-	// {
-	// 	id: 15,
-	// 	name: 'jwt',
-	// 	image: jwt,
-	// },
-];
