+++ app/constants/providers.js
@@ -1,0 +1,51 @@
+const google = require('assets/providers/google.png');
+const facebook = require('assets/providers/facebook.png');
+const apple = require('assets/providers/apple.png');
+const discord = require('assets/providers/discord.png');
+const github = require('assets/providers/github.png');
+const kakao = require('assets/providers/kakao.png');
+const linkedin = require('assets/providers/linkedin.png');
+const twitter = require('assets/providers/twitter.png');
+
+export const providers = [
+	{
+		id: 1,
+		name: 'google',
+		image: google,
+	},
+	{
+		id: 2,
+		name: 'facebook',
+		image: facebook,
+	},
+	{
+		id: 3,
+		name: 'discord',
+		image: discord,
+	},
+	{
+		id: 4,
+		name: 'apple',
+		image: apple,
+	},
+	{
+		id: 5,
+		name: 'github',
+		image: github,
+	},
+	{
+		id: 6,
+		name: 'kakao',
+		image: kakao,
+	},
+	{
+		id: 7,
+		name: 'linkedin',
+		image: linkedin,
+	},
+	{
+		id: 8,
+		name: 'twitter',
+		image: twitter,
+	},
+];
